feat: add pay controller

+++ src/main/java/com/example/orderapi/service/pointhistory/PointHistoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.orderapi.service.pointhistory;
 
 import com.example.orderapi.dto.pointhistory.PointHistoryCreateRequest;
 import com.example.orderapi.dto.pointhistory.PointHistoryResponse;
+import com.example.orderapi.entity.pointhistory.HistoryTypes;
 import com.example.orderapi.entity.pointhistory.PointHistory;
 import com.example.orderapi.entity.pointpolicy.PointPolicy;
 import com.example.orderapi.exception.notfound.PointHistoryNotFoundException;
@@ -10,11 +11,13 @@ import com.example.orderapi.repository.pointhistory.PointHistoryRepository;
 import com.example.orderapi.repository.pointpolicy.PointPolicyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -23,13 +26,13 @@ public class PointHistoryServiceImpl implements PointHistoryService {
     private final PointHistoryRepository pointHistoryRepository;
     private final PointPolicyRepository pointPolicyRepository;
 
-
     @Override
     public Page<PointHistoryResponse> getPointHistoriesByMemberId(Long memberId, Pageable pageable) {
         Page<PointHistory> histories = pointHistoryRepository.findAllByMemberId(memberId, pageable);
-        if(histories.isEmpty()){
-            throw new PointHistoryNotFoundException(String.format("memberId=%d's pointHistory is null", memberId));        }
 
+        if (histories.isEmpty()) {
+            throw new PointHistoryNotFoundException(String.format("No point histories found for memberId=%d", memberId));
+        }
 
         return histories.map(PointHistoryResponse::fromEntity);
     }
@@ -37,14 +40,20 @@ public class PointHistoryServiceImpl implements PointHistoryService {
     @Override
     public Page<PointHistoryResponse> getPointHistories(Pageable pageable) {
         Page<PointHistory> histories = pointHistoryRepository.findAll(pageable);
-        if(histories.isEmpty()){
-            throw new PointHistoryNotFoundException("point history is not found");
+
+        if (histories.isEmpty()) {
+            throw new PointHistoryNotFoundException("No point histories found.");
         }
+
         return histories.map(PointHistoryResponse::fromEntity);
     }
 
     @Override
     public void removePointHistoryById(Long pointHistoryId) {
+        if (!pointHistoryRepository.existsById(pointHistoryId)) {
+            throw new PointHistoryNotFoundException(String.format("PointHistory with id=%d not found", pointHistoryId));
+        }
+
         pointHistoryRepository.deleteById(pointHistoryId);
     }
 
@@ -55,34 +64,59 @@ public class PointHistoryServiceImpl implements PointHistoryService {
 
     @Override
     public PointHistoryResponse getPointHistory(Long pointHistoryId) {
-        PointHistory history = pointHistoryRepository.findById(pointHistoryId).orElse(null);
-        if(Objects.isNull(history)){
-            throw new PointHistoryNotFoundException(String.format("pointHistoryId=%d is not found", pointHistoryId));        }
+        PointHistory history = pointHistoryRepository.findById(pointHistoryId)
+                .orElseThrow(() -> new PointHistoryNotFoundException(
+                        String.format("PointHistory with id=%d not found", pointHistoryId)
+                ));
 
         return PointHistoryResponse.fromEntity(history);
     }
 
     @Override
-    public PointHistoryResponse createPointHistory(PointHistoryCreateRequest request) {
-        PointPolicy pointPolicy = pointPolicyRepository.findById(request.getPointPolicyId()).orElse(null);
-        if(Objects.isNull(pointPolicy)){
-            throw new PointPolicyNotFoundException(String.format("pointPolicyId=%d is not found", request.getPointPolicyId()));        }
+    public PointHistoryResponse createPointHistory(Long memberId, PointHistoryCreateRequest request) {
+        PointPolicy pointPolicy = pointPolicyRepository.findById(request.getPointPolicyId())
+                .orElseThrow(() -> new PointPolicyNotFoundException(
+                        String.format("PointPolicy with id=%d not found", request.getPointPolicyId())
+                ));
 
-        PointHistory pointHistory = new PointHistory(null,
+        PointHistory pointHistory = PointHistory.ofCreate(
                 request.getTypes(),
                 pointPolicy.getAmount(),
-                LocalDateTime.now(),
                 pointPolicy.getName(),
-                request.getMemberId());
+                memberId
+        );
+
         PointHistory savedHistory = pointHistoryRepository.save(pointHistory);
         return PointHistoryResponse.fromEntity(savedHistory);
     }
 
-
     @Override
-    public Integer getTotalPointByMemberId(Long pointId) {
-        return pointHistoryRepository.sumAmount(pointId);
+    public Integer getTotalPointByMemberId(Long memberId) {
+        return pointHistoryRepository.sumAmount(memberId);
     }
 
+    public Page<PointHistoryResponse> getPointHistoriesWithinPeriod(Long memberId, LocalDate startDate, LocalDate endDate,String sortDirection, Pageable pageable) {
+        LocalDateTime startDateTime = startDate.atStartOfDay();         // 시작일 00:00:00
+        LocalDateTime endDateTime = endDate.plusDays(1).atStartOfDay(); // 종료일의 다음날 00:00:00
+
+        Sort.Direction direction = Sort.Direction.fromString(sortDirection.toUpperCase());
+        Pageable sortedPageable = PageRequest.of(pageable.getPageNumber(), pageable.getPageSize(), Sort.by(direction, "changedAt"));
+
+        Page<PointHistory> histories = pointHistoryRepository.findAllByChangedAtBetween(memberId, startDateTime, endDateTime, sortedPageable);
+        if (histories.isEmpty()) {
+            throw new PointHistoryNotFoundException("No point histories found.");
+        }
+
+        return histories.map(PointHistoryResponse::fromEntity);
+    }
+
+    @Override
+    public Page<PointHistoryResponse> getPointHistoriesWithState(Long memberId, HistoryTypes state, Pageable pageable) {
+        Page<PointHistory> histories = pointHistoryRepository.findAllByMemberIdAndTypes(memberId, state, pageable);
+        if (histories.isEmpty()) {
+            throw new PointHistoryNotFoundException("No point histories found.");
+        }
+        return histories.map(PointHistoryResponse::fromEntity);
+    }
 
 }
